+++ opal/mca/pmix/pmix1xx/pmix/include/pmix/pmix_common.h.in
@@ -183,7 +183,7 @@ BEGIN_C_DECLS
 
 /****    PMIX ERROR CONSTANTS    ****/
 /* PMIx errors are always negative, with 0 reserved for success */
-#define PMIX_ERROR_MIN  -41  // set equal to number of non-zero entries in enum
+#define PMIX_ERROR_MIN  -42  // set equal to number of non-zero entries in enum
 
 typedef enum {
     PMIX_ERR_UNPACK_READ_PAST_END_OF_BUFFER = PMIX_ERROR_MIN,
@@ -230,6 +230,7 @@ typedef enum {
     PMIX_ERR_INVALID_CRED,
     PMIX_EXISTS,
 
+    PMIX_ERR_SILENT,
     PMIX_ERROR,
     PMIX_SUCCESS
 } pmix_status_t;
