+++ opal/mca/pmix/pmix1xx/pmix/src/server/Makefile.am
@@ -16,4 +16,5 @@ sources += \
         src/server/pmix_server.c \
         src/server/pmix_server_ops.c \
         src/server/pmix_server_regex.c \
-        src/server/pmix_server_listener.c
+        src/server/pmix_server_listener.c \
+        src/server/pmix_server_get.c
