+++ opal/mca/pmix/pmix1xx/pmix/src/server/pmix_server_ops.c
@@ -58,246 +58,6 @@
 
 pmix_server_module_t pmix_host_server = {0};
 
-static void dmdx_cbfunc(pmix_status_t status, const char *data,
-                        size_t ndata, void *cbdata,
-                        pmix_release_cbfunc_t relfn, void *relcbdata);
-
-typedef struct {
-    pmix_object_t super;
-    pmix_event_t ev;
-    pmix_status_t status;
-    const char *data;
-    size_t ndata;
-    pmix_dmdx_local_t *lcd;
-    pmix_release_cbfunc_t relcbfunc;
-    void *cbdata;
-} pmix_dmdx_reply_caddy_t;
-PMIX_CLASS_INSTANCE(pmix_dmdx_reply_caddy_t,
-                   pmix_object_t, NULL, NULL);
-
-
-static void relfn(void *cbdata)
-{
-    char *data = (char*)cbdata;
-    free(data);
-}
-
-static pmix_status_t _satisfy_request(pmix_nspace_t *nptr, int rank,
-                                         pmix_hash_table_t *ht,
-                                         pmix_modex_cbfunc_t cbfunc, void *cbdata)
-{
-    pmix_status_t rc;
-    pmix_buffer_t pbkt, xfer;
-    pmix_value_t *val;
-    char *data;
-    size_t sz;
-
-    /* check to see if this data already has been
-     * obtained as a result of a prior direct modex request from
-     * another local peer */
-    rc = pmix_hash_fetch(ht, rank, "modex", &val);
-    if (PMIX_SUCCESS == rc && NULL != val) {
-        PMIX_CONSTRUCT(&pbkt, pmix_buffer_t);
-        PMIX_CONSTRUCT(&xfer, pmix_buffer_t);
-        pmix_buffer_t *pxfer = &xfer;
-        PMIX_LOAD_BUFFER(&xfer, val->data.bo.bytes, val->data.bo.size);
-        pmix_bfrop.pack(&pbkt, &pxfer, 1, PMIX_BUFFER);
-        xfer.base_ptr = NULL;
-        xfer.bytes_used = 0;
-        PMIX_DESTRUCT(&xfer);
-        PMIX_VALUE_RELEASE(val);
-        PMIX_UNLOAD_BUFFER(&pbkt, data, sz);
-        PMIX_DESTRUCT(&pbkt);
-        /* pass it back */
-        cbfunc(rc, data, sz, cbdata, relfn, data);
-        return rc;
-    }
-    return PMIX_ERR_NOT_FOUND;
-}
-
-pmix_status_t pmix_pending_request(pmix_nspace_t *nptr, int rank,
-                                   pmix_info_t *info, size_t ninfo,
-                                   pmix_modex_cbfunc_t cbfunc, void *cbdata)
-{
-    pmix_dmdx_local_t *lcd = NULL, *cd;
-    pmix_rank_info_t *iptr;
-    pmix_hash_table_t *ht;
-    pmix_status_t rc;
-
-    /* 1. Try to satisfy the request right now */
-
-    /* by default we are looking for the remote data */
-    ht = &nptr->server->remote;
-    PMIX_LIST_FOREACH(iptr, &nptr->server->ranks, pmix_rank_info_t) {
-        if (iptr->rank == rank) {
-            /* in case it is known local rank - check local table */
-            ht = &nptr->server->mylocal;
-            break;
-        }
-    }
-
-    rc = _satisfy_request(nptr, rank, ht, cbfunc, cbdata);
-    if( PMIX_SUCCESS == rc ){
-        /* request was successfully satisfied */
-        PMIX_INFO_FREE(info, ninfo);
-        return rc;
-    }
-
-    /* 2. We were unable to satisfy request right now.
-     * Look for existing requests to this namespace/rank */
-    PMIX_LIST_FOREACH(cd, &pmix_server_globals.local_reqs, pmix_dmdx_local_t) {
-        if (0 != strncmp(nptr->nspace, cd->proc.nspace, PMIX_MAX_NSLEN) ||
-                rank != cd->proc.rank ) {
-            continue;
-        }
-        lcd = cd;
-        break;
-    }
-
-    /* 3. If no requests exists then:
-     * - if all local clients are registered then we were called because
-     * the remote data was requested. Create request and call direct modex
-     * to retrieve the data
-     * - if not all local ranks were registered, we need to wait untill
-     * pmix_pending_localy_fin would be called to resolve this. Just add the
-     * request for now.
-     */
-    if (NULL == lcd) {
-        lcd = PMIX_NEW(pmix_dmdx_local_t);
-        if (NULL == lcd){
-            PMIX_INFO_FREE(info, ninfo);
-            return PMIX_ERR_NOMEM;
-        }
-        strncpy(lcd->proc.nspace, nptr->nspace, PMIX_MAX_NSLEN);
-        lcd->proc.rank = rank;
-        lcd->info = info;
-        lcd->ninfo = ninfo;
-        pmix_list_append(&pmix_server_globals.local_reqs, &lcd->super);
-
-        /* check & send request if need/possible */
-        if (nptr->server->all_registered && NULL == info) {
-            if (NULL != pmix_host_server.direct_modex) {
-                pmix_host_server.direct_modex(&lcd->proc, info, ninfo, dmdx_cbfunc, lcd);
-            } else {
-                /* if we don't have direct modex feature, just respond with "not found" */
-                cbfunc(PMIX_ERR_NOT_FOUND, NULL, 0, cbdata, NULL, NULL);
-                PMIX_INFO_FREE(info, ninfo);
-                pmix_list_remove_item(&pmix_server_globals.local_reqs, &lcd->super);
-                PMIX_LIST_DESTRUCT(&lcd->loc_reqs);
-                PMIX_RELEASE(lcd);
-                return PMIX_SUCCESS;
-            }
-        }
-    }
-    pmix_dmdx_request_t *req = PMIX_NEW(pmix_dmdx_request_t);
-    req->cbfunc = cbfunc;
-    req->cbdata = cbdata;
-    pmix_list_append(&lcd->loc_reqs, &req->super);
-    return PMIX_SUCCESS;
-}
-
-void pmix_pending_nspace_fix(pmix_nspace_t *nptr)
-{
-    pmix_dmdx_local_t *cd, *cd_next;
-
-    /* Now when we know all local ranks, go along request list and ask for remote data
-     * for the non-local ranks, and resolve all pending requests for local procs
-     * that were waiting for registration to complete
-     */
-    PMIX_LIST_FOREACH_SAFE(cd, cd_next, &pmix_server_globals.local_reqs, pmix_dmdx_local_t) {
-        pmix_rank_info_t *info;
-        bool found = false;
-
-        if (0 != strncmp(nptr->nspace, cd->proc.nspace, PMIX_MAX_NSLEN) ) {
-            continue;
-        }
-
-        PMIX_LIST_FOREACH(info, &nptr->server->ranks, pmix_rank_info_t) {
-            if (info->rank == cd->proc.rank) {
-                found = true;
-                break;
-            }
-        }
-
-        /* if not found - this is remote process and we need to send
-         * corresponding direct modex request */
-        if( !found ){
-            if( NULL != pmix_host_server.direct_modex ){
-                pmix_host_server.direct_modex(&cd->proc, cd->info, cd->ninfo, dmdx_cbfunc, cd);
-            } else {
-                pmix_dmdx_request_t *req, *req_next;
-                PMIX_LIST_FOREACH_SAFE(req, req_next, &cd->loc_reqs, pmix_dmdx_request_t) {
-                    req->cbfunc(PMIX_ERR_NOT_FOUND, NULL, 0, req->cbdata, NULL, NULL);
-                    pmix_list_remove_item(&cd->loc_reqs, &req->super);
-                    PMIX_RELEASE(req);
-                }
-                pmix_list_remove_item(&pmix_server_globals.local_reqs, &cd->super);
-                PMIX_RELEASE(cd);
-            }
-        }
-    }
-}
-
-/* Resolve pending requests to this namespace/rank */
-pmix_status_t pmix_pending_resolve(pmix_nspace_t *nptr, int rank,
-                                   pmix_status_t status, pmix_dmdx_local_t *lcd)
-{
-    pmix_dmdx_local_t *cd;
-
-    /* find corresponding request (if exists) */
-    if( NULL == lcd ){
-        PMIX_LIST_FOREACH(cd, &pmix_server_globals.local_reqs, pmix_dmdx_local_t) {
-            if (0 != strncmp(nptr->nspace, cd->proc.nspace, PMIX_MAX_NSLEN) ||
-                    rank != cd->proc.rank) {
-                continue;
-            }
-            lcd = cd;
-            break;
-        }
-    }
-
-    /* If somebody was interested in this rank */
-    if( NULL != lcd ){
-        pmix_dmdx_request_t *req;
-
-        if (PMIX_SUCCESS != status){
-            /* if we've got an error for this request - just forward it*/
-            PMIX_LIST_FOREACH(req, &lcd->loc_reqs, pmix_dmdx_request_t) {
-                /* if we can't satisfy this request - respond with error */
-                req->cbfunc(status, NULL, 0, req->cbdata, NULL, NULL);
-            }
-        } else {
-            /* if we've got the blob - try to satisfy requests */
-            pmix_hash_table_t *ht;
-            pmix_rank_info_t *iptr;
-
-            /* by default we are looking for the remote data */
-            ht = &nptr->server->remote;
-            /* check if this rank is local */
-            PMIX_LIST_FOREACH(iptr, &nptr->server->ranks, pmix_rank_info_t) {
-                if (iptr->rank == rank) {
-                    ht = &nptr->server->mylocal;
-                    break;
-                }
-            }
-
-            /* run through all the requests to this rank */
-            PMIX_LIST_FOREACH(req, &lcd->loc_reqs, pmix_dmdx_request_t) {
-                pmix_status_t rc;
-                rc = _satisfy_request(nptr, rank, ht, req->cbfunc, req->cbdata);
-                if( PMIX_SUCCESS != rc ){
-                    /* if we can't satisfy this particular request (missing key?) */
-                    req->cbfunc(rc, NULL, 0, req->cbdata, NULL, NULL);
-                }
-            }
-        }
-        /* remove all requests to this rank and cleanup the corresponding structure */
-        pmix_list_remove_item(&pmix_server_globals.local_reqs, (pmix_list_item_t*)lcd);
-        PMIX_RELEASE(lcd);
-    }
-    return PMIX_SUCCESS;
-}
-
 pmix_status_t pmix_server_abort(pmix_peer_t *peer, pmix_buffer_t *buf,
                                 pmix_op_cbfunc_t cbfunc, void *cbdata)
 {
@@ -436,13 +196,7 @@ pmix_status_t pmix_server_commit(pmix_peer_t *peer, pmix_buffer_t *buf)
              * may not be a contribution */
             if (PMIX_SUCCESS == pmix_hash_fetch(&nptr->server->myremote, info->rank, "modex", &val) &&
                 NULL != val) {
-                PMIX_CONSTRUCT(&xfer, pmix_buffer_t);
-                PMIX_LOAD_BUFFER(&xfer, val->data.bo.bytes, val->data.bo.size);
-                pmix_buffer_t *pxfer = &xfer;
-                pmix_bfrop.pack(&pbkt, &pxfer, 1, PMIX_BUFFER);
-                xfer.base_ptr = NULL;
-                xfer.bytes_used = 0;
-                PMIX_DESTRUCT(&xfer);
+                PMIX_LOAD_BUFFER(&pbkt, val->data.bo.bytes, val->data.bo.size);
                 PMIX_VALUE_RELEASE(val);
             }
             PMIX_UNLOAD_BUFFER(&pbkt, data, sz);
@@ -457,7 +211,7 @@ pmix_status_t pmix_server_commit(pmix_peer_t *peer, pmix_buffer_t *buf)
             PMIX_RELEASE(dcd);
         }
     }
-    /* see if anyone local is waiting on this data- could be more than one */
+    /* see if anyone local is waiting on this data - could be more than one */
     return pmix_pending_resolve(nptr, info->rank, PMIX_SUCCESS, NULL);
 }
 
@@ -824,163 +578,6 @@ pmix_status_t pmix_server_fence(pmix_server_caddy_t *cd,
  cleanup:
     PMIX_PROC_FREE(procs, nprocs);
     return rc;
-}
-
-static void _process_dmdx_reply(int fd, short args, void *cbdata)
-{
-    pmix_dmdx_reply_caddy_t *caddy = (pmix_dmdx_reply_caddy_t *)cbdata;
-    pmix_kval_t *kp;
-    pmix_nspace_t *ns, *nptr;
-    pmix_status_t rc;
-
-    pmix_output_verbose(2, pmix_globals.debug_output,
-                    "[%s:%d] queue dmdx reply from %s:%d",
-                    __FILE__, __LINE__,
-                    caddy->lcd->proc.nspace, caddy->lcd->proc.rank);
-
-    /* find the nspace object for this client */
-    nptr = NULL;
-    PMIX_LIST_FOREACH(ns, &pmix_globals.nspaces, pmix_nspace_t) {
-        if (0 == strcmp(caddy->lcd->proc.nspace, ns->nspace)) {
-            nptr = ns;
-            break;
-        }
-    }
-
-    if (NULL == nptr) {
-        /* should be impossible */
-        PMIX_ERROR_LOG(PMIX_ERR_NOT_FOUND);
-        caddy->status = PMIX_ERR_NOT_FOUND;
-        goto cleanup;
-    }
-
-    if (PMIX_SUCCESS == caddy->status) {
-        kp = PMIX_NEW(pmix_kval_t);
-        kp->key = strdup("modex");
-        PMIX_VALUE_CREATE(kp->value, 1);
-        kp->value->type = PMIX_BYTE_OBJECT;
-        /* we don't know if the host is going to save this data
-         * or not, so we have to copy it */
-        kp->value->data.bo.bytes = (char*)malloc(caddy->ndata);
-        memcpy(kp->value->data.bo.bytes, caddy->data, caddy->ndata);
-        kp->value->data.bo.size = caddy->ndata;
-        /* store it in the appropriate hash */
-        if (PMIX_SUCCESS != (rc = pmix_hash_store(&nptr->server->remote, caddy->lcd->proc.rank, kp))) {
-            PMIX_ERROR_LOG(rc);
-        }
-        PMIX_RELEASE(kp);  // maintain acctg
-    }
-
-cleanup:
-    /* always execute the callback to avoid having the client hang */
-    pmix_pending_resolve(nptr, caddy->lcd->proc.rank, caddy->status, caddy->lcd);
-
-    /* now call the release function so the host server
-     * knows it can release the data */
-    if (NULL != caddy->relcbfunc) {
-        caddy->relcbfunc(caddy->cbdata);
-    }
-    PMIX_RELEASE(caddy);
-}
-
-static void dmdx_cbfunc(pmix_status_t status,
-                        const char *data, size_t ndata, void *cbdata,
-                        pmix_release_cbfunc_t release_fn, void *release_cbdata)
-{
-    pmix_dmdx_reply_caddy_t *caddy;
-    caddy = PMIX_NEW(pmix_dmdx_reply_caddy_t);
-    caddy->status = status;
-    /* point to the callers cbfunc */
-    caddy->relcbfunc = release_fn;
-    caddy->cbdata = release_cbdata;
-
-    caddy->data   = data;
-    caddy->ndata  = ndata;
-    caddy->lcd    = (pmix_dmdx_local_t *)cbdata;
-    pmix_output_verbose(2, pmix_globals.debug_output, "[%s:%d] queue dmdx reply %s:%d",
-                        __FILE__, __LINE__,
-                        caddy->lcd->proc.nspace, caddy->lcd->proc.rank);
-    event_assign(&caddy->ev, pmix_globals.evbase, -1, EV_WRITE,
-                 _process_dmdx_reply, caddy);
-    event_priority_set(&caddy->ev, 0);
-    event_active(&caddy->ev, EV_WRITE, 1);
-}
-
-pmix_status_t pmix_server_get(pmix_buffer_t *buf,
-                              pmix_modex_cbfunc_t cbfunc,
-                              void *cbdata)
-{
-    int32_t cnt;
-    pmix_status_t rc;
-    int rank;
-    char *cptr;
-    char nspace[PMIX_MAX_NSLEN+1];
-    pmix_nspace_t *ns, *nptr;
-    pmix_info_t *info=NULL;
-    size_t ninfo=0;
-
-    pmix_output_verbose(2, pmix_globals.debug_output,
-                        "recvd GET");
-
-    /* setup */
-    memset(nspace, 0, sizeof(nspace));
-
-    /* retrieve the nspace and rank of the requested proc */
-    cnt = 1;
-    if (PMIX_SUCCESS != (rc = pmix_bfrop.unpack(buf, &cptr, &cnt, PMIX_STRING))) {
-        PMIX_ERROR_LOG(rc);
-        return rc;
-    }
-    (void)strncpy(nspace, cptr, PMIX_MAX_NSLEN);
-    free(cptr);
-    cnt = 1;
-    if (PMIX_SUCCESS != (rc = pmix_bfrop.unpack(buf, &rank, &cnt, PMIX_INT))) {
-        PMIX_ERROR_LOG(rc);
-        return rc;
-    }
-    /* find the nspace object for this client */
-    nptr = NULL;
-    PMIX_LIST_FOREACH(ns, &pmix_globals.nspaces, pmix_nspace_t) {
-        if (0 == strcmp(nspace, ns->nspace)) {
-            nptr = ns;
-            break;
-        }
-    }
-
-    pmix_output_verbose(2, pmix_globals.debug_output,
-                        "%s:%d EXECUTE GET FOR %s:%d",
-                        pmix_globals.myid.nspace,
-                        pmix_globals.myid.rank, nspace, rank);
-
-    /* retrieve any provided info structs */
-    cnt = 1;
-    if (PMIX_SUCCESS != (rc = pmix_bfrop.unpack(buf, &ninfo, &cnt, PMIX_SIZE))) {
-        PMIX_ERROR_LOG(rc);
-        return rc;
-    }
-    if (0 < ninfo) {
-        PMIX_INFO_CREATE(info, ninfo);
-        cnt = ninfo;
-        if (PMIX_SUCCESS != (rc = pmix_bfrop.unpack(buf, info, &cnt, PMIX_INFO))) {
-            PMIX_ERROR_LOG(rc);
-            PMIX_INFO_FREE(info, ninfo);
-            return rc;
-        }
-    }
-
-    if (NULL == nptr) {
-        /* this is for an nspace we don't know about yet, so
-         * give the host server a chance to tell us about it */
-        nptr = PMIX_NEW(pmix_nspace_t);
-        (void)strncpy(nptr->nspace, nspace, PMIX_MAX_NSLEN);
-        pmix_list_append(&pmix_globals.nspaces, &nptr->super);
-    }
-    /* if we don't have any ranks for this job, protect ourselves here */
-    if (NULL == nptr->server) {
-        nptr->server = PMIX_NEW(pmix_server_nspace_t);
-    }
-
-    return pmix_pending_request(nptr, rank, info, ninfo, cbfunc, cbdata);
 }
 
 pmix_status_t pmix_server_publish(pmix_peer_t *peer,
